Add files via upload
In the windows the user can calculate the resistance, capacitance, and time constant of each relaxation and the reconstruction of Z*/M* data of each electrochemical process or multiple processes. Finally, it is also possible to perform subtractions by selecting a range of peaks from the DFRT and to generate the simulated impedance dispersion.